simulator: Fix joystick button polling
There's BUTTON_PRESSED(which_button, down_latches) which
returns true if the button is pressed, but clears the bit,
and won't return true again until the button is pressed again,
and then there's button_poll(which_button), which returns
true if the button is pressed and doesn't clear a latch so
you can use it for continuous action by holding down a button.
The joystick code didn't behave this way.  Now it does.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>