goodbye-gulag: smooth out wall-gliding at corners
When travelling parallel to a wall steering slightly into the wall
gliding along it, and coming to an outside corner, the collision
detection would stop you at the corner because it would think
you collided with the perpendicular wall.


    +-------
    |
   X|
    |

If the player is where the 'X' is, travelling north-northeast, he would get
stopped at the corner (the '+') because the collision detection would think
he got stopped by the horzontal wall.

Fix this by testing against each wall as if they were slightly shorter
than they actually are.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>