simulator: simulate infrared communication between badges via UDP
This worked in 2020, but I guess it got lost when the
simulator got significantly revamped in 2022.

The 2020 version would also simulate IR over a serial port.
I don't remember why.  I think there might have been a way
to have the simulator interact with a real badge via the
serial port as if it were using the IR, but I'm not really
sure.  Anyway, I'm not re-implementing that here.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>