menus: fix bug in menus
If you got onto the Thursday schedule menu,
then scrolled backwards around the menu a few
times, then selected "Lunch", address sanitizer
would abort the program.

The problem is when we push a new menu, we need to
set menu_scroll_start_item to 0 so we start off at
the top of the new menu.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>