Enable alternate method of displaying menus
Basically, replace display_menu() function with indirection
through a function pointer, and there's a new item under
"Settings" for "Menu Style" which allows selecting either
"Legacy Menus" or "New Menus" which set this function pointer
to one of the two options.

For now, the two styles are identical, but the new one has
a red border instead of green/cyan border.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>