Do not diplay "back" items in new menu
They aren't needed in the new menu system.

Also, in the simulator, detect infinite loops in menus (e.g. menus
with all SKIP and HIDDEN items, so you can't advance to the "next"
or "previous" item because there is no item that isn't skipped
or hidden.)

Also, modify menu navigation code to skip BACK items if the new
menus are being used.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>