Redesign how menu animations work
We can't just do the entire animation within a single call
to new_display_menu().  Now new_display_menu() doesn't
actually draw anything.  Intead it returns what the new
menu selection is (as always) but sets up an animation
state record and sets a flag indicating that an animation
is in progress.  menus() then draws one frame of this
animation by calling animate_menu() if this animation
in progress flag is set.  At the end of the animation,
the flag is unset. This means that between each frame of
the animation, we return all the way out back to the
main loop in ProcessIO().

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>