new-badge-monsters: reprocess images to save space
Reprocess the badge monster image files with png-to-badge-asset
which should reduce flash memory usage of the image colors by 1/3
compared to the old method. In this case it doesn't do much though
because all the badge monster images are 4 bit images, and each
image has only a handful of colors (at most 16), so we might save
something like 16 bytes per image.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>