menu-system: Make it not crash coming back from screensaver
If you let the screensaver activate while the menu is up, then
press two directional buttons, e.g. left, left, or left, right, etc.
the first press would work ok, but the 2nd would crash. (See issue #12
in HackRVA/badge2024 project.)

This change prevents it from crashing.  It's not perfect though, because
it puts the menu back on the first item of the main menu, which is probably
not the correct item (but it's better than NULL.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>