clue: Use Fisher-Yates shuffling algorithm
Not that it matters much, but the naive method of shuffling
I was using previously (swap every card in the deck with another
card at a randomly chosen position in the deck) lands on some
permutations more often than others.  E.g. in a 3 card deck,
there would be 9 outcomes of the shuffle, but there are only 6
permutations of 3 cards, and 9 is not divisible by 6, thus some
permutations must be chosen more often than others.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>