cmake: define PICO_XOSC_STARTUP_DELAY_MULTIPLIER=64 to solve "no flash retain" bug
A common issue with the RP2040 is sensitivity to crystals which take
longer to stabilize. This macro increases the stabilization period to
ensure the pico does not get stuck starting the crystal on boot. This
appears to be the main culprit on many "no flash retain" units; I have
tested two units which experience the bug on previous firmware that
launch consistently after this change.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>