clue: do not respond to IR packets too quickly.
When transmitting an IR packet the procedure is as follows
(see ir_rp2040.c:ir_send_complete_message()):

1. Turn of IR interrupt
2. send IR packet
3. Wait at least 200 ms (maybe more)
4. Turn on IR interrupt.

This means if you design an IR protocol with query and response,
at least 200 ms must pass between transmission of the query and
the start of the response, or else the response will not be
received.

The fix is to ensure we do not transmit the response without first
waiting 400ms after the query is received, and 3000ms between
sending queries.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>